requirements.txt
asgiref==3.3.3
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.11
django-allauth==0.49.0
django-mock-queries==2.1.7
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==13.3.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
#MarkupSafe==2.1.0
mock==4.0.3
model-bakery==1.4.0
oauthlib==3.2.0
packaging==21.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.1.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.0.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.2.0
